Some work on componentizing the minimized chats UI
