+++ spec/chatbox.js
@@ -193,82 +193,6 @@ describe("Chatboxes", function () {
             done();
         }));
 
-        it("can be trimmed to conserve space",
-            mock.initConverse(['rosterGroupsFetched'], {},
-            async function (done, _converse) {
-
-            spyOn(_converse.chatboxviews, 'trimChats');
-
-            const trimmed_chatboxes = _converse.minimized_chats;
-            spyOn(trimmed_chatboxes, 'addChat').and.callThrough();
-            spyOn(trimmed_chatboxes, 'removeChat').and.callThrough();
-
-            await mock.waitForRoster(_converse, 'current');
-            await mock.openControlBox(_converse);
-            expect(_converse.chatboxviews.trimChats.calls.count()).toBe(1);
-
-            let jid, chatboxview;
-            // openControlBox was called earlier, so the controlbox is
-            // visible, but no other chat boxes have been created.
-            expect(_converse.chatboxes.length).toEqual(1);
-            expect(document.querySelectorAll("#conversejs .chatbox").length).toBe(1); // Controlbox is open
-
-            _converse.rosterview.update(); // XXX: Hack to make sure $roster element is attached.
-            await u.waitUntil(() => _converse.rosterview.el.querySelectorAll('.roster-group li').length);
-            // Test that they can be maximized again
-            const online_contacts = _converse.rosterview.el.querySelectorAll('.roster-group .current-xmpp-contact a.open-chat');
-            expect(online_contacts.length).toBe(17);
-            let i;
-            for (i=0; i<online_contacts.length; i++) {
-                const el = online_contacts[i];
-                el.click();
-            }
-            await u.waitUntil(() => _converse.chatboxes.length == 16);
-            expect(_converse.chatboxviews.trimChats.calls.count()).toBe(16);
-
-            _converse.api.chatviews.get().forEach(v => spyOn(v, 'onMinimized').and.callThrough());
-            for (i=0; i<online_contacts.length; i++) {
-                const el = online_contacts[i];
-                jid = _.trim(el.textContent.trim()).replace(/ /g,'.').toLowerCase() + '@montague.lit';
-                chatboxview = _converse.chatboxviews.get(jid);
-                chatboxview.model.set({'minimized': true});
-                expect(trimmed_chatboxes.addChat).toHaveBeenCalled();
-                expect(chatboxview.onMinimized).toHaveBeenCalled();
-            }
-            await u.waitUntil(() => _converse.chatboxviews.keys().length);
-            var key = _converse.chatboxviews.keys()[1];
-            const trimmedview = trimmed_chatboxes.get(key);
-            const chatbox = trimmedview.model;
-            spyOn(chatbox, 'maximize').and.callThrough();
-            spyOn(trimmedview, 'restore').and.callThrough();
-            trimmedview.delegateEvents();
-            trimmedview.el.querySelector("a.restore-chat").click();
-
-            expect(trimmedview.restore).toHaveBeenCalled();
-            expect(chatbox.maximize).toHaveBeenCalled();
-            expect(_converse.chatboxviews.trimChats.calls.count()).toBe(17);
-            done();
-        }));
-
-        it("can be opened in minimized mode initially",
-            mock.initConverse(
-                ['rosterGroupsFetched'], {},
-                async function (done, _converse) {
-
-            await mock.waitForRoster(_converse, 'current');
-            const sender_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
-            await _converse.api.chats.create(sender_jid, {'minimized': true});
-            await u.waitUntil(() => _converse.chatboxes.length > 1);
-            const chatBoxView = _converse.chatboxviews.get(sender_jid);
-            expect(u.isVisible(chatBoxView.el)).toBeFalsy();
-
-            const minimized_chat = _converse.minimized_chats.get(sender_jid);
-            expect(minimized_chat).toBeTruthy();
-            expect(u.isVisible(minimized_chat.el)).toBeTruthy();
-            done();
-        }));
-
-
         it("is focused if its already open and you click on its corresponding roster item",
             mock.initConverse(['rosterGroupsFetched', 'chatBoxesFetched'], {},
                 async function (done, _converse) {
@@ -361,42 +285,6 @@ describe("Chatboxes", function () {
             expect(chatview.close).toHaveBeenCalled();
             await new Promise(resolve => _converse.api.listen.once('chatBoxClosed', resolve));
             expect(_converse.api.trigger).toHaveBeenCalledWith('chatBoxClosed', jasmine.any(Object));
-            done();
-        }));
-
-        it("can be minimized by clicking a DOM element with class 'toggle-chatbox-button'",
-            mock.initConverse(
-                ['rosterGroupsFetched', 'chatBoxesFetched'], {},
-                async function (done, _converse) {
-
-            await mock.waitForRoster(_converse, 'current');
-            await mock.openControlBox(_converse);
-
-            const contact_jid = mock.cur_names[7].replace(/ /g,'.').toLowerCase() + '@montague.lit';
-            await u.waitUntil(() => _converse.rosterview.el.querySelectorAll('.roster-group').length);
-            await mock.openChatBoxFor(_converse, contact_jid);
-            const trimmed_chatboxes = _converse.minimized_chats;
-            const chatview = _converse.chatboxviews.get(contact_jid);
-            spyOn(chatview, 'minimize').and.callThrough();
-            spyOn(_converse.api, "trigger").and.callThrough();
-            // We need to rebind all events otherwise our spy won't be called
-            chatview.delegateEvents();
-            chatview.el.querySelector('.toggle-chatbox-button').click();
-
-            expect(chatview.minimize).toHaveBeenCalled();
-            expect(_converse.api.trigger).toHaveBeenCalledWith('chatBoxMinimized', jasmine.any(Object));
-            expect(_converse.api.trigger.calls.count(), 2);
-            expect(u.isVisible(chatview.el)).toBeFalsy();
-            expect(chatview.model.get('minimized')).toBeTruthy();
-            chatview.el.querySelector('.toggle-chatbox-button').click();
-            const trimmedview = trimmed_chatboxes.get(chatview.model.get('id'));
-            spyOn(trimmedview, 'restore').and.callThrough();
-            trimmedview.delegateEvents();
-            trimmedview.el.querySelector("a.restore-chat").click();
-
-            expect(trimmedview.restore).toHaveBeenCalled();
-            expect(_converse.api.trigger).toHaveBeenCalledWith('chatBoxMaximized', jasmine.any(Object));
-            expect(chatview.model.get('minimized')).toBeFalsy();
             done();
         }));
 
@@ -1588,84 +1476,6 @@ describe("Chatboxes", function () {
             expect(select_msgs_indicator().textContent).toBe('1');
             await mock.openChatBoxFor(_converse, sender_jid);
             expect(select_msgs_indicator().textContent).toBe('1');
-            done();
-        }));
-    });
-
-    describe("A Minimized ChatBoxView's Unread Message Count", function () {
-
-        it("is displayed when scrolled up chatbox is minimized after receiving unread messages",
-            mock.initConverse(
-                ['rosterGroupsFetched', 'chatBoxesFetched'], {},
-                async function (done, _converse) {
-
-            await mock.waitForRoster(_converse, 'current', 1);
-            const sender_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
-            await mock.openChatBoxFor(_converse, sender_jid);
-            const msgFactory = function () {
-                return mock.createChatMessage(_converse, sender_jid, 'This message will be received as unread, but eventually will be read');
-            };
-            const selectUnreadMsgCount = function () {
-                const minimizedChatBoxView = _converse.minimized_chats.get(sender_jid);
-                return minimizedChatBoxView.el.querySelector('.message-count');
-            };
-            const chatbox = _converse.chatboxes.get(sender_jid);
-            chatbox.save('scrolled', true);
-            _converse.handleMessageStanza(msgFactory());
-            await u.waitUntil(() => chatbox.messages.length);
-            const chatboxview = _converse.chatboxviews.get(sender_jid);
-            chatboxview.minimize();
-
-            const unread_count = selectUnreadMsgCount();
-            expect(u.isVisible(unread_count)).toBeTruthy();
-            expect(unread_count.innerHTML.replace(/<!---->/g, '')).toBe('1');
-            done();
-        }));
-
-        it("is incremented when message is received and windows is not focused",
-            mock.initConverse(
-                ['rosterGroupsFetched', 'chatBoxesFetched'], {},
-                async function (done, _converse) {
-
-            await mock.waitForRoster(_converse, 'current', 1);
-            const sender_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
-            const view = await mock.openChatBoxFor(_converse, sender_jid)
-            const msgFactory = function () {
-                return mock.createChatMessage(_converse, sender_jid,
-                    'This message will be received as unread, but eventually will be read');
-            };
-            const selectUnreadMsgCount = function () {
-                const minimizedChatBoxView = _converse.minimized_chats.get(sender_jid);
-                return minimizedChatBoxView.el.querySelector('.message-count');
-            };
-            view.minimize();
-            _converse.handleMessageStanza(msgFactory());
-            await u.waitUntil(() => view.model.messages.length);
-            const unread_count = selectUnreadMsgCount();
-            expect(u.isVisible(unread_count)).toBeTruthy();
-            expect(unread_count.innerHTML.replace(/<!---->/g, '')).toBe('1');
-            done();
-        }));
-
-        it("will render Openstreetmap-URL from geo-URI",
-            mock.initConverse(
-                ['rosterGroupsFetched', 'chatBoxesFetched'], {},
-                async function (done, _converse) {
-
-            await mock.waitForRoster(_converse, 'current', 1);
-
-            const message = "geo:37.786971,-122.399677";
-            const contact_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
-            await mock.openChatBoxFor(_converse, contact_jid);
-            const view = _converse.chatboxviews.get(contact_jid);
-            spyOn(view.model, 'sendMessage').and.callThrough();
-            mock.sendMessage(view, message);
-            await u.waitUntil(() => view.el.querySelectorAll('.chat-content .chat-msg').length, 1000);
-            expect(view.model.sendMessage).toHaveBeenCalled();
-            const msg = sizzle('.chat-content .chat-msg:last .chat-msg__text', view.el).pop();
-            await u.waitUntil(() => msg.innerHTML.replace(/\<!----\>/g, '') ===
-                '<a target="_blank" rel="noopener" href="https://www.openstreetmap.org/?mlat=37.786971&amp;'+
-                'mlon=-122.399677#map=18/37.786971/-122.399677">https://www.openstreetmap.org/?mlat=37.786971&amp;mlon=-122.399677#map=18/37.786971/-122.399677</a>');
             done();
         }));
     });
