--- src/templates/chats_panel.html
@@ -1,2 +1,0 @@
-<a id="toggle-minimized-chats" href="#" class="row no-gutters"></a>
-<div class="flyout minimized-chats-flyout row no-gutters"></div>
