+++ webpack.html
@@ -20,6 +20,7 @@
         }
     });
     converse.initialize({
+        // root: new DocumentFragment(),
         show_send_button: true,
         auto_away: 300,
         auto_register_muc_nickname: true,
